requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
msgspec==0.18.6
python-slugify==8.0.4
requests==2.31.0
text-unidecode==1.3
Unidecode==1.3.8
urllib3==2.2.1
